wpt: Have `ServoRefTestExecutor` inherit from `RefTestExecutor`
Since `ProcessTestExecutor` is only used by Servo and doesn't add much
on top of `ServoExecutor`, merge the former into the latter.

Also make `ServoExecutor` a mixin so that the derived classes can
inherit from upstream `wptrunner`'s classes and avoid the diamond
inheritance of `TestExecutor` base class.

Another issue is that new changes in the
`RefTestImplementation.get_screenshot_list` method upstream conflicts
with the assumptions in Servo's executor:
  - The `dpi` argument is used as an `int` although it is a string
  - The `viewport_size` argument is treated as a tuple of ints but it is
    in fact a string of the format `WxH`. It also treats the viewport
    dimensions as being specified in device pixel coordinates while
    Servo treats it as logical coordinates and will scale it by `dpi`.

These issues need to discussed with upstream and patched in
`RefTestImplementation`.

Finally, add back `**kwargs` argument to `ServoCrashtestExecutor`'s
constructor to fix #40322.

Fixes #40288, #40322
